Fix tests for parallel execution
- Remove debug printfs
- Fix warnings on data types in unittests
- Fix Makefile for unittests to hard-code serial MPI
- Bug-Fix in examples/fortran/dejong.f: This was trying to store an
  8-byte pointer into a 4-byte integer on 64bit architectures
- Fixes for printf statements: Only print if MPI rank 0

+++ examples/fortran/dejong.f
@@ -251,7 +251,7 @@ c
       double precision  Lower(5), Upper(5)
       common /data/     BinLen, NumCoords, Lower, Upper, gray_on
       double precision  value
-      integer comm
+      integer(8) comm
 
 
       comm = PGAGetCommunicator(ctx)
